Refactor CLI imports
