+++ zenml/cli/step.py
@@ -17,7 +17,7 @@ from tabulate import tabulate
 
 from zenml.cli.cli import cli
 from zenml.cli.utils import pass_repo
-from zenml.core.repo import Repository
+from zenml.repo import Repository
 
 
 @cli.group()
